Beatmap writer

+++ OsuParsers/Beatmaps/Sections/DifficultySection.cs
@@ -6,7 +6,7 @@ namespace OsuParsers.Beatmaps.Sections
         public float CircleSize { get; set; }
         public float OverallDifficulty { get; set; }
         public float ApproachRate { get; set; }
-        public float SliderMultiplier { get; set; }
-        public float SliderTickRate { get; set; }
+        public double SliderMultiplier { get; set; }
+        public double SliderTickRate { get; set; }
     }
 }
